+++ src/main/java/com/example/glow/entity/mySql/ContractCloseDtResource.java
@@ -25,7 +25,7 @@ public class ContractCloseDtResource implements Serializable {
 
     @Id
     @Column(name = "CONTRACT_ID")
-    private Integer contractId;
+    private Long contractId;
 
     @Column(name = "CONTRACT_CLOSE_DT")
     private Date contractCloseDt;
