+++ src/main/java/com/example/glow/entity/mySql/InsuranceProduct2GroupResource.java
@@ -23,7 +23,7 @@ public class InsuranceProduct2GroupResource {
 
     @Id
     @Column(name = "INSURANCE_PRODUCT_ID")
-    private Integer insuranceProductId;
+    private Long insuranceProductId;
 
     @Column(name = "INSURANCE_PRODUCT_2_GROUP")
     private String insuranceProduct2Group;
