+++ src/main/java/com/example/glow/entity/mySql/IsutanceContractXClientResource.java
@@ -23,10 +23,10 @@ public class IsutanceContractXClientResource {
 
     @Id
     @Column(name = "CONTRACT_ID")
-    private Integer contractId;
+    private Long contractId;
 
     @Column(name = "CLIENT_ID")
-    private Integer clientId;
+    private Long clientId;
 
     @Column(name = "EFFECTIVE_FROM_DTTM")
     private Timestamp effectiveFromDttm;
